WIP: chore(js) Finish clean settings components
