+++ js/packages/berty-storybook/shared-components/Footer.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { TouchableOpacity, View, Image } from 'react-native'
+import { TouchableOpacity, View, Image, StyleSheet } from 'react-native'
 import { Icon } from 'react-native-ui-kitten'
 import { colors, styles } from '../styles'
 
@@ -23,6 +23,14 @@ type FooterProps = {
 	center: ButtonFooterProps
 	right: ButtonFooterProps
 }
+
+// Styles
+const _footerStyles = StyleSheet.create({
+	backgroundBlur: {
+		height: 100,
+		opacity: 0.8,
+	},
+})
 
 const ButtonFooter: React.FC<ButtonFooterProps> = ({
 	backgroundColor = colors.white,
@@ -65,22 +73,34 @@ const ButtonFooter: React.FC<ButtonFooterProps> = ({
 )
 
 export const Footer: React.FC<FooterProps> = ({ left, center, right }) => (
-	<View
-		style={[
-			styles.absolute,
-			styles.bottom,
-			styles.left,
-			styles.right,
-			styles.marginBottom,
-			styles.padding,
-			styles.row,
-			styles.spaceAround,
-			styles.alignItems,
-		]}
-	>
-		<ButtonFooter {...left} />
-		<ButtonFooter {...center} />
-		<ButtonFooter {...right} />
+	<View>
+		<View
+			style={[
+				styles.bgWhite,
+				styles.absolute,
+				styles.bottom,
+				styles.left,
+				styles.right,
+				_footerStyles.backgroundBlur,
+			]}
+		/>
+		<View
+			style={[
+				styles.absolute,
+				styles.bottom,
+				styles.left,
+				styles.right,
+				styles.marginBottom,
+				styles.padding,
+				styles.row,
+				styles.spaceAround,
+				styles.alignItems,
+			]}
+		>
+			<ButtonFooter {...left} />
+			<ButtonFooter {...center} />
+			<ButtonFooter {...right} />
+		</View>
 	</View>
 )
