+++ js/packages/berty-storybook/shared-components/SettingsButtons.tsx
@@ -1,14 +1,8 @@
-import React, { useState, useEffect } from 'react'
-import {
-	View,
-	TouchableOpacity,
-	StyleSheet,
-	LayoutChangeEvent,
-	SafeAreaView,
-	StyleProp,
-} from 'react-native'
+import React, { useState } from 'react'
+import { View, TouchableOpacity, StyleSheet, StyleProp } from 'react-native'
 import { Text, Icon, Toggle } from 'react-native-ui-kitten'
 import { styles, colors } from '../styles'
+import { CircleAvatar } from './CircleAvatar'
 
 //
 // Button Setting
@@ -16,11 +10,12 @@ import { styles, colors } from '../styles'
 
 // Type
 type SettingButtonProps = {
+	name: string
+	image?: string
 	icon?: string
 	iconSize?: number
 	iconColor?: string
 	iconDependToggle?: boolean
-	name: string
 	children?: React.ReactNode
 	state?: {
 		value: string
@@ -39,8 +34,10 @@ type SettingButtonProps = {
 	actionIconColor?: string
 	actionToggle?: React.Dispatch<React.SetStateAction<any>>
 	varToggle?: boolean
+	style?: StyleProp<any>[]
 	// action
 	previewValue?: string
+	previewValueColor?: string
 }
 
 // Style
@@ -57,6 +54,7 @@ const _stylesSettingButton = StyleSheet.create({
 
 export const ButtonSetting: React.FC<SettingButtonProps> = ({
 	name,
+	image = null,
 	icon = null,
 	iconSize = 30,
 	iconColor = colors.blue,
@@ -68,20 +66,13 @@ export const ButtonSetting: React.FC<SettingButtonProps> = ({
 	actionToggle = null,
 	varToggle = null,
 	previewValue = null,
+	previewValueColor = colors.black,
 	alone = true,
 	toggled = false,
+	style = null,
 	actionIcon = !toggled && 'arrow-ios-forward',
 }) => {
-	const [layout, setLayout] = useState()
 	const [isToggle, setIsToggle] = useState()
-
-	const handleLayout = (e: LayoutChangeEvent) => {
-		setLayout(e.nativeEvent.layout.width)
-	}
-
-	const handleChangeToggle = () => {
-		setIsToggle(!isToggle)
-	}
 
 	return (
 		<TouchableOpacity
@@ -89,6 +80,7 @@ export const ButtonSetting: React.FC<SettingButtonProps> = ({
 			style={[
 				_stylesSettingButton.settingButton,
 				styles.bgWhite,
+				style,
 				alone ? styles.borderRadius : null,
 				alone ? styles.shadow : null,
 				alone ? { marginTop: 20 } : null,
@@ -121,6 +113,11 @@ export const ButtonSetting: React.FC<SettingButtonProps> = ({
 							/>
 						</View>
 					)}
+					{image && (
+						<View>
+							<CircleAvatar avatarUri={image} withCircle={false} size={35} />
+						</View>
+					)}
 					<View>
 						<Text
 							style={[styles.fontFamily, styles.littlePaddingLeft, { color: colors.black }]}
@@ -130,10 +127,7 @@ export const ButtonSetting: React.FC<SettingButtonProps> = ({
 						</Text>
 					</View>
 				</View>
-				<View
-					onLayout={(e) => handleLayout(e)}
-					style={[styles.row, styles.center, styles.alignItems]}
-				>
+				<View style={[styles.row, styles.center, styles.alignItems]}>
 					{state && state.value && state.color && state.bgColor && (
 						<View style={[styles.row, styles.marginRight, styles.alignItems]}>
 							{state && state.icon && (
@@ -178,7 +172,8 @@ export const ButtonSetting: React.FC<SettingButtonProps> = ({
 								style={[
 									styles.fontFamily,
 									styles.littlePaddingRight,
-									{ fontWeight: 'bold', color: colors.black },
+									styles.textBold,
+									{ color: previewValueColor },
 								]}
 								category='s4'
 							>
@@ -222,17 +217,27 @@ type FactionButtonSettingProps = {
 	icon?: string
 	iconSize?: number
 	iconColor?: string
+	state?: {
+		value: string
+		color: string
+		bgColor: string
+		icon?: string
+		iconSize?: number
+		iconColor?: string
+		stateIcon?: string
+		stateIconColor?: string
+	}
 	style?: StyleProp<any>
 }
 
 // Styles
-
 export const FactionButtonSetting: React.FC<FactionButtonSettingProps> = ({
 	children,
 	name = null,
 	icon = null,
-	iconSize = 0,
-	iconColor = null,
+	iconSize = 30,
+	iconColor = colors.blue,
+	state = {},
 	style = null,
 }) => (
 	<View
@@ -240,8 +245,8 @@ export const FactionButtonSetting: React.FC<FactionButtonSettingProps> = ({
 			styles.bgWhite,
 			styles.shadow,
 			styles.borderRadius,
-			styles.littlePaddingLeft,
-			styles.littlePaddingRight,
+			styles.paddingLeft,
+			styles.paddingRight,
 			style,
 		]}
 	>
@@ -254,13 +259,48 @@ export const FactionButtonSetting: React.FC<FactionButtonSettingProps> = ({
 						</View>
 					)}
 					<View>
-						<Text
-							style={[styles.fontFamily, styles.littlePaddingLeft, { fontWeight: 'bold' }]}
-							category='s4'
-						>
+						<Text style={[styles.fontFamily, styles.littlePaddingLeft]} category='s4'>
 							{name}
 						</Text>
 					</View>
+					{state && state.value && state.color && state.bgColor && (
+						<View style={[styles.rowRev, styles.marginRight, styles.alignItems, styles.flex]}>
+							{state && state.icon && (
+								<Icon
+									style={[styles.littleMarginRight]}
+									name={state.icon}
+									width={state.iconSize}
+									height={state.iconSize}
+									fill={state.iconColor}
+								/>
+							)}
+							<View
+								style={[
+									styles.row,
+									styles.spaceEvenly,
+									styles.alignItems,
+									styles.borderRadius,
+									{ backgroundColor: state.bgColor },
+									_stylesSettingButton.statePaddingBox,
+								]}
+							>
+								{state.stateIcon && (
+									<Icon
+										style={[styles.center, { marginRight: 5 }]}
+										name={state.stateIcon}
+										width={13}
+										height={13}
+										fill={state.stateIconColor}
+									/>
+								)}
+								<Text
+									style={[styles.center, { color: state.color, fontSize: 8, fontWeight: 'bold' }]}
+								>
+									{state.value}
+								</Text>
+							</View>
+						</View>
+					)}
 				</View>
 				<View
 					style={[
@@ -293,124 +333,109 @@ export const FactionButtonSetting: React.FC<FactionButtonSettingProps> = ({
 )
 
 //
-// Header Info Settings
+// ButtonSettingRow => The tree buttons in many settings screens (chat and settings)
 //
 
-// Type
-type HeaderInfoSettingsProps = {
-	children: React.ReactNode
-	bgColor?: string
+// Types
+type ButtonSettingRowProps = {
+	state: {
+		name: string
+		icon: string
+		color: string
+		style: StyleProp<any>
+	}[]
+	numberOfLines?: number
+	style?: StyleProp<any>
+	styleText?: StyleProp<any>
 }
 
-// Style
-const _stylesHeaderInfoSettings = StyleSheet.create({})
+// Styles
+const _buttonSettingRowStyles = StyleSheet.create({
+	textPadding: {
+		paddingTop: 6,
+	},
+})
 
-export const HeaderInfoSettings: React.FC<HeaderInfoSettingsProps> = ({ children, bgColor }) => (
+export const ButtonSettingRow: React.FC<ButtonSettingRowProps> = ({
+	state,
+	numberOfLines = 1,
+	style = null,
+	styleText = null,
+}) => (
 	<View
 		style={[
-			{ backgroundColor: 'rgba(206,210,255,0.3)' },
-			styles.marginLeft,
-			styles.marginRight,
-			styles.borderRadius,
-			styles.padding,
+			styles.flex,
+			styles.row,
+			styles.spaceBetween,
+			styles.alignItems,
 			styles.marginTop,
+			style,
 		]}
 	>
-		<View style={[styles.justifyContent]}>{children}</View>
+		{state.map((obj) => (
+			<TouchableOpacity
+				style={[
+					styles.flex,
+					styles.padding,
+					styles.alignItems,
+					styles.spaceCenter,
+					styles.shadow,
+					styles.bgWhite,
+					styles.borderRadius,
+					obj.style,
+				]}
+			>
+				<Icon name={obj.icon} width={30} height={30} fill={obj.color} />
+				<Text
+					style={[
+						styles.textCenter,
+						styles.fontFamily,
+						styles.textBlack,
+						styleText,
+						_buttonSettingRowStyles.textPadding,
+					]}
+					category='s4'
+					numberOfLines={numberOfLines}
+				>
+					{obj.name}
+				</Text>
+			</TouchableOpacity>
+		))}
 	</View>
 )
 
 //
-// Header Settings
+// ButtonSettingItem
 //
 
-// Type
-type HeaderSettingsProps = {
-	title: string
-	children?: React.ReactNode
-	// titleSize?: number
-	// titleColor?: string
-	bgColor?: string
-	undo?: boolean
-	undoIcon?: string
-	undoIconSize?: number
-	undoIconColor?: string
-	desc?: string
-	descFontSize?: number
-	descColor?: string
-	action?: React.Dispatch<any>
-	actionValue?: boolean
-	actionIcon?: string
-	actionIconSize?: number
-	actionIconColor?: string
+// Types
+type ButtonSettingItem = {
+	value: string
+	color?: string
+	icon?: string
+	iconSize?: number
+	iconColor?: string
 }
 
-export const HeaderSettings: React.FC<HeaderSettingsProps> = ({
-	title,
-	children = null,
-	bgColor = colors.blue,
-	undo = true,
-	undoIcon = 'arrow-back-outline',
-	undoIconSize = 30,
-	undoIconColor = colors.white,
-	desc = null,
-	descFontSize = 10,
-	descColor = colors.white,
-	action = null,
-	actionValue = null,
-	actionIcon = 'swap',
-	actionIconSize = 30,
-	actionIconColor = colors.white,
-}) => (
-	<SafeAreaView
-		style={[
-			styles.flex,
-			styles.borderBottomLeftRadius,
-			styles.borderBottomRightRadius,
-			{ backgroundColor: bgColor },
-		]}
-	>
-		<View style={[styles.padding]}>
-			<View style={[styles.row, styles.justifyContent]}>
-				{undo && (
-					<TouchableOpacity style={[styles.flex, styles.start, styles.center]}>
-						<Icon name={undoIcon} width={undoIconSize} height={undoIconSize} fill={undoIconColor} />
-					</TouchableOpacity>
-				)}
-				<View style={[styles.center]}>
-					<Text category='h3' style={[styles.fontFamily, styles.textWhite, styles.center]}>
-						{title}
-					</Text>
-				</View>
-				<View style={[styles.flex, styles.end, styles.center]}>
-					{action && (
-						<TouchableOpacity style={[styles.end]} onPress={() => action(!actionValue)}>
-							<Icon
-								name={actionIcon}
-								width={actionIconSize}
-								height={actionIconSize}
-								fill={actionIconColor}
-							/>
-						</TouchableOpacity>
-					)}
-				</View>
-			</View>
-			{desc && (
-				<Text
-					style={[
-						styles.center,
-						{ fontSize: descFontSize, color: descColor },
-						styles.textCenter,
-						styles.bigPaddingLeft,
-						styles.bigPaddingRight,
-					]}
-				>
-					{desc}
-				</Text>
-			)}
-			{children}
-		</View>
-	</SafeAreaView>
-)
+// Styles
+const _buttonSettingItemStyles = StyleSheet.create({
+	updateFeatureText: {
+		fontSize: 11,
+		paddingLeft: 8,
+	},
+})
 
-export default ButtonSetting
+export const ButtonSettingItem: React.FC<ButtonSettingItem> = ({
+	value,
+	color = colors.white,
+	icon = 'checkmark-circle-2',
+	iconSize = 12,
+	iconColor = colors.lightBlue,
+}) => (
+	<View style={[styles.row, styles.littlePaddingLeft, styles.alignItems]}>
+		<Icon name={icon} width={iconSize} height={iconSize} fill={iconColor} />
+		<Text style={[styles.textBold, _buttonSettingItemStyles.updateFeatureText, { color }]}>
+			{value}
+		</Text>
+	</View>
+)
