+++ js/packages/berty-storybook/shared-props/User.tsx
@@ -2,3 +2,7 @@ export type UserProps = {
 	avatarUri: string
 	name: string
 }
+
+export type RequestProps = {
+	user: UserProps
+}
